# tests/__snapshots__/openapi.core.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecisionRequestDto": {
        "properties": {
          "decision": {
            "type": "string"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "scope": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "ApprovalGateDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "decider_agent": {
            "type": [
              "null",
              "string"
            ]
          },
          "evidence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          },
          "gate_index": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "gate_kind": {
            "type": "string"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "gate_index",
          "gate_kind",
          "status",
          "decider_agent",
          "reason",
          "evidence",
          "created_at",
          "decided_at"
        ],
        "type": "object"
      },
      "ApprovalGatesDto": {
        "properties": {
          "approval_id": {
            "format": "uuid",
            "type": "string"
          },
          "gates": {
            "items": {
              "$ref": "#/components/schemas/ApprovalGateDto"
            },
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "run_id",
          "status",
          "gates"
        ],
        "type": "object"
      },
...
      "JsonElement": {},
...
      "ModelRouteCandidateDto": {
        "properties": {
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "provider_instance_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ModelRouteWriteRequestDto": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ModelRouteCandidateDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "PreAuthorizationRequestDto": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "lane_key": {
            "type": [
              "null",
              "string"
            ]
          },
          "max_uses": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "parameter_constraint_hash": {
            "type": [
              "null",
              "string"
            ]
          },
          "risk_max": {
            "type": [
              "null",
              "string"
            ]
          },
          "run_id": {
            "format": "uuid",
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "tool_scope": {
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          }
        },
        "required": [
          "run_id",
          "lane_key",
          "tool_scope",
          "parameter_constraint_hash",
          "risk_max",
          "max_uses",
          "expires_at",
          "summary"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TinadecCore.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/approvals": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
    "/api/v1/approvals/pre-authorizations": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreAuthorizationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
    "/api/v1/approvals/{approvalId}/gates": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "approvalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalGatesDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
    "/api/v1/approvals/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
    "/api/v1/approvals/{id}/decision": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/model-providers/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
    "/api/v1/model-providers/{id}/models/refresh": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/model-routes/{purpose}": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "purpose",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRouteWriteRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/prompt-fragments/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/prompt-fragments/{id}/effectiveness": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}